allow arbitrary style of ordered lists: https://github.com/metanorma/metanorma-iso/issues/1139
