+++ Gemfile.devel
@@ -1,2 +1,2 @@
-gem "metanorma-standoc", git: "https://github.com/metanorma/metanorma-standoc", branch: "fix/primary-docidentifier"
+gem "metanorma-standoc", git: "https://github.com/metanorma/metanorma-standoc", branch: "main"
 gem "isodoc", git: "https://github.com/metanorma/isodoc", branch: "main"
