+++ lib/metanorma/iso/isodoc.rng
@@ -438,28 +438,6 @@
         </optional>
       </element>
     </define>
-    <define name="FigureAttributes">
-      <attribute name="id">
-        <data type="ID"/>
-      </attribute>
-      <optional>
-        <attribute name="unnumbered">
-          <data type="boolean"/>
-        </attribute>
-      </optional>
-      <optional>
-        <attribute name="number"/>
-      </optional>
-      <optional>
-        <attribute name="subsequence"/>
-      </optional>
-      <optional>
-        <attribute name="class"/>
-      </optional>
-      <optional>
-        <attribute name="width"/>
-      </optional>
-    </define>
     <define name="source">
       <element name="source">
         <attribute name="status">
@@ -2532,6 +2510,28 @@
     </optional>
     <optional>
       <attribute name="columns"/>
+    </optional>
+  </define>
+  <define name="FigureAttributes">
+    <attribute name="id">
+      <data type="ID"/>
+    </attribute>
+    <optional>
+      <attribute name="unnumbered">
+        <data type="boolean"/>
+      </attribute>
+    </optional>
+    <optional>
+      <attribute name="number"/>
+    </optional>
+    <optional>
+      <attribute name="subsequence"/>
+    </optional>
+    <optional>
+      <attribute name="class"/>
+    </optional>
+    <optional>
+      <attribute name="width"/>
     </optional>
   </define>
   <start>
