+++ spec/isodoc/terms_spec.rb
@@ -91,7 +91,7 @@ RSpec.describe IsoDoc do
                 </ul>
               </termexample>
               <termsource status="modified">[SOURCE: <origin bibitemid="ISO7301" type="inline" citeas="ISO 7301:2011"><locality type="clause"><referenceFrom>3.1</referenceFrom></locality><span class="stdpublisher">ISO </span><span class="stddocNumber">7301</span>:<span class="stdyear">2011</span>, <span class="citesec">3.1</span></origin>, modified
-                    –
+                    &#x2014;
                    The term "cargo rice" is shown as deprecated, and Note 1 to entry is not included here]</termsource>
             </term>
             <term id='paddy'><name>1.2</name>
@@ -141,7 +141,7 @@ RSpec.describe IsoDoc do
     html = <<~OUTPUT
       #{HTML_HDR}
                    <div id="_"><h1>1  Terms and Definitions</h1><p class="TermNum" id="paddy1">1.1</p><p class="Terms" style="text-align:left;"><b>paddy</b></p><p id="_"><rice> rice retaining its husk after threshing</p><div id="_" class="example"><p><span class="example_label">EXAMPLE 1</span>  Foreign seeds, husks, bran, sand, dust.</p><ul><li>A</li></ul></div><div id="_" class="example"><p><span class="example_label">EXAMPLE 2</span>  </p><ul><li>A</li></ul></div><p>[SOURCE: <span class="stdpublisher">ISO </span><span class="stddocNumber">7301</span>:<span class="stdyear">2011</span>, <span class="citesec">3.1</span>, modified
-                     –
+                     &#x2014;
                     The term "cargo rice" is shown as deprecated, and Note 1 to entry is not included here]</p><p class="TermNum" id="paddy">1.2</p><p class="Terms" style="text-align:left;"><b>paddy</b></p><p class="AltTerms" style="text-align:left;">paddy rice</p><p class="AltTerms" style="text-align:left;">rough rice</p><p class="DeprecatedTerms" style="text-align:left;">DEPRECATED: cargo rice</p><p id="_">rice retaining its husk after threshing</p><div id="_" class="Note"><p>Note 1 to entry: The starch of waxy rice consists almost entirely of amylopectin. The
                    kernels have a tendency to stick together after cooking.</p></div><div id="_" class="Note"><p>Note 2 to entry: <ul><li>A</li></ul><p id="_">The starch of waxy rice consists almost entirely of amylopectin. The
                    kernels have a tendency to stick together after cooking.</p></p></div><div id="_" class="example"><p><span class="example_label">EXAMPLE</span>  </p><ul><li>A</li></ul></div><p>[SOURCE: <span class="stdpublisher">ISO </span><span class="stddocNumber">7301</span>:<span class="stdyear">2011</span>, <span class="citesec">3.1</span>]</p><p class="TermNum" id="A">1.3</p><p class="Terms" style="text-align:left;"><b>term1</b></p>
@@ -157,7 +157,7 @@ RSpec.describe IsoDoc do
 
     word = <<~OUTPUT
            <div id="_"><h1>1<span style="mso-tab-count:1">  </span>Terms and Definitions</h1><p class="TermNum" id="paddy1">1.1</p><p class="Terms" style="text-align:left;"><b>paddy</b></p><p class="Definition" id="_"><rice> rice retaining its husk after threshing</p><div id="_" class="example"><p><span class="example_label">EXAMPLE 1</span><span style="mso-tab-count:1">  </span>Foreign seeds, husks, bran, sand, dust.</p><ul><li>A</li></ul></div><div id="_" class="example"><p><span class="example_label">EXAMPLE 2</span><span style="mso-tab-count:1">  </span></p><ul><li>A</li></ul></div><p class="Source">[SOURCE: <span class="stdpublisher">ISO </span><span class="stddocNumber">7301</span>:<span class="stdyear">2011</span>, <span class="citesec">3.1</span>, modified
-                     –
+                     &#x2014;
                     The term "cargo rice" is shown as deprecated, and Note 1 to entry is not included here]</p><p class="TermNum" id="paddy">1.2</p><p class="Terms" style="text-align:left;"><b>paddy</b></p><p class="AltTerms" style="text-align:left;">paddy rice</p><p class="AltTerms" style="text-align:left;">rough rice</p><p class="DeprecatedTerms" style="text-align:left;">DEPRECATED: cargo rice</p><p class="Definition" id="_">rice retaining its husk after threshing</p><div id="_" class="Note"><p class="Note">Note 1 to entry: The starch of waxy rice consists almost entirely of amylopectin. The
                    kernels have a tendency to stick together after cooking.</p></div><div id="_" class="Note"><p class="Note">Note 2 to entry: <ul><li>A</li></ul><p id="_">The starch of waxy rice consists almost entirely of amylopectin. The
                    kernels have a tendency to stick together after cooking.</p></p></div><div id="_" class="example"><p><span class="example_label">EXAMPLE</span><span style="mso-tab-count:1">  </span></p><ul><li>A</li></ul></div><p class="Source">[SOURCE: <span class="stdpublisher">ISO </span><span class="stddocNumber">7301</span>:<span class="stdyear">2011</span>, <span class="citesec">3.1</span>]</p><p class="TermNum" id="A">1.3</p><p class="Terms" style="text-align:left;"><b>term1</b></p>
@@ -305,17 +305,17 @@ RSpec.describe IsoDoc do
             <p id="_eb29b35e-123e-4d1c-b50b-2714d41e747f">rice retaining its husk after threshing</p>
           </definition>
           <termsource status="identical">[SOURCE: <origin citeas=""><termref base="IEV" target="xyz">t1</termref></origin>
-           –
+           &#x2014;
             with adjustments
 
         ;
           <origin citeas=""><termref base="IEV" target="xyz"/></origin>, modified
-           –
+           &#x2014;
             with adjustments
 
         ;
           <origin citeas=""><termref base="IEV" target="xyz"/></origin>, modified
-           –
+           &#x2014;
             with adjustments
 
         ;
